+++ src/Silk.Core.Data/GuildContext.cs
@@ -28,6 +28,7 @@ namespace Silk.Core.Data
         public DbSet<Tag> Tags { get; set; } = null!;
 
         public DbSet<Reminder> Reminders { get; set; } = null!;
+        public DbSet<GuildModConfig> GuildModConfigs { get; set; }
 
         protected override void OnModelCreating(ModelBuilder builder)
         {
