+++ src/Silk.Core.Data/MediatR/Guilds/Config/Mod/GetGuildModConfigRequest.cs
@@ -1,0 +1,19 @@
+﻿using System.Threading;
+using System.Threading.Tasks;
+using MediatR;
+using Microsoft.EntityFrameworkCore;
+using Silk.Core.Data.Models;
+
+namespace Silk.Core.Data.MediatR.Guilds.Config
+{
+	public sealed record GetGuildModConfigRequest(ulong guildId) : IRequest<GuildModConfig>;
+	
+	public sealed class GetGuildModConfigHandler : IRequestHandler<GetGuildModConfigRequest, GuildModConfig>
+	{
+		private readonly GuildContext _db;
+		public GetGuildModConfigHandler(GuildContext db) => _db = db;
+
+		public Task<GuildModConfig> Handle(GetGuildModConfigRequest request, CancellationToken cancellationToken)
+			=> _db.GuildModConfigs.FirstOrDefaultAsync(c => c.GuildId == request.guildId, cancellationToken);
+	}
+}
