--- src/Silk.Core.Data/MediatR/Guilds/UpdateGuildConfigRequest.cs
@@ -1,93 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Threading;
-using System.Threading.Tasks;
-using MediatR;
-using Microsoft.EntityFrameworkCore;
-using Silk.Core.Data.Models;
-
-namespace Silk.Core.Data.MediatR.Guilds
-{
-    /// <summary>
-    ///     Request for updating a <see cref="GuildConfig" /> for a Guild.
-    ///     <remarks>
-    ///         When it comes to collections, such as <see cref="SelfAssignableRoles" /> only the *added* elements should be passed to this request,
-    ///         else they'll be removed.
-    ///     </remarks>
-    /// </summary>
-    /// <param name="GuildId">The Id of the Guild</param>
-    public record UpdateGuildConfigRequest(ulong GuildId) : IRequest<GuildConfig?>
-    {
-        public ulong? MuteRoleId { get; init; }
-        public ulong? LoggingChannel { get; init; }
-        public ulong? GreetingChannelId { get; init; }
-        public ulong? VerificationRoleId { get; init; }
-
-        public bool? ScanInvites { get; init; }
-        public GreetingOption? GreetingOption { get; init; }
-        public bool? BlacklistWords { get; init; }
-        public bool? BlacklistInvites { get; init; }
-        public bool? LogMembersJoining { get; init; }
-        public bool? UseAggressiveRegex { get; init; }
-        public bool? WarnOnMatchedInvite { get; init; }
-        public bool? DeleteOnMatchedInvite { get; init; }
-        public bool? GreetOnVerificationRole { get; init; }
-        public bool? GreetOnScreeningComplete { get; init; }
-
-        public int? MaxUserMentions { get; init; }
-        public int? MaxRoleMentions { get; init; }
-
-        public string? GreetingText { get; init; }
-
-        public List<Invite>? AllowedInvites { get; init; }
-        public List<DisabledCommand>? DisabledCommands { get; init; }
-
-        public List<InfractionStep>? InfractionSteps { get; init; }
-        //public List<BlacklistedWord>? BlacklistedWords { get; init; }
-    }
-
-    /// <summary>
-    ///     The default handler for <see cref="UpdateGuildConfigRequest" />.
-    /// </summary>
-    public class UpdateGuildConfigHandler : IRequestHandler<UpdateGuildConfigRequest, GuildConfig?>
-    {
-        private readonly GuildContext _db;
-
-        public UpdateGuildConfigHandler(GuildContext db) => _db = db;
-
-        public async Task<GuildConfig?> Handle(UpdateGuildConfigRequest request, CancellationToken cancellationToken)
-        {
-            GuildConfig config = await _db.GuildConfigs
-                .Include(c => c.InfractionSteps)
-                .AsSplitQuery()
-                .FirstOrDefaultAsync(g => g.GuildId == request.GuildId, cancellationToken);
-
-            config.MuteRoleId = request.MuteRoleId ?? config.MuteRoleId;
-            config.GreetingOption = request.GreetingOption ?? config.GreetingOption;
-            config.LoggingChannel = request.LoggingChannel ?? config.LoggingChannel;
-            config.GreetingChannel = request.GreetingChannelId ?? config.GreetingChannel;
-            config.VerificationRole = request.VerificationRoleId ?? config.VerificationRole;
-
-            config.ScanInvites = request.ScanInvites ?? config.ScanInvites;
-            config.BlacklistWords = request.BlacklistWords ?? config.BlacklistWords;
-            config.BlacklistInvites = request.BlacklistInvites ?? config.BlacklistInvites;
-            config.LogMemberJoins = request.LogMembersJoining ?? config.LogMemberJoins;
-            config.UseAggressiveRegex = request.UseAggressiveRegex ?? config.UseAggressiveRegex;
-            config.WarnOnMatchedInvite = request.WarnOnMatchedInvite ?? config.WarnOnMatchedInvite;
-            config.DeleteMessageOnMatchedInvite = request.DeleteOnMatchedInvite ?? config.DeleteMessageOnMatchedInvite;
-            config.GreetOnVerificationRole = request.GreetOnVerificationRole ?? config.GreetOnVerificationRole;
-            config.GreetOnScreeningComplete = request.GreetOnScreeningComplete ?? config.GreetOnVerificationRole;
-
-            config.MaxUserMentions = request.MaxUserMentions ?? config.MaxUserMentions;
-            config.MaxRoleMentions = request.MaxRoleMentions ?? config.MaxRoleMentions;
-
-            config.GreetingText = request.GreetingText ?? config.GreetingText;
-
-            config.InfractionSteps = request.InfractionSteps ?? config.InfractionSteps;
-            config.AllowedInvites = request.AllowedInvites ?? config.AllowedInvites;
-            config.DisabledCommands = request.DisabledCommands ?? config.DisabledCommands;
-            
-            await _db.SaveChangesAsync(cancellationToken);
-            return config;
-        }
-    }
-}
