+++ src/Silk.Core.Data/Models/Guild/Guild.cs
@@ -11,6 +11,7 @@ namespace Silk.Core.Data.Models
         [StringLength(5)]
         public string Prefix { get; set; } = "";
         public GuildConfig Configuration { get; set; } = new();
+        public GuildModConfig ModConfig { get; set; } = new();
         public List<User> Users { get; set; } = new();
         public List<Infraction> Infractions { get; set; } = new();
         public List<Tag> Tags { get; set; } = new();
