+++ src/Silk.Core.Data/Silk.Core.Data.csproj
@@ -22,8 +22,4 @@
         <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="5.0.6" />
     </ItemGroup>
 
-    <ItemGroup>
-      <Folder Include="Models\Guild\Config" />
-    </ItemGroup>
-
 </Project>
