+++ src/Silk.Core/Commands/Moderation/KickCommand.cs
@@ -4,7 +4,6 @@ using DSharpPlus;
 using DSharpPlus.CommandsNext;
 using DSharpPlus.CommandsNext.Attributes;
 using DSharpPlus.Entities;
-using Microsoft.Extensions.Logging;
 using Silk.Core.Data.Models;
 using Silk.Core.Services.Interfaces;
 using Silk.Core.Types;
@@ -18,16 +17,10 @@ namespace Silk.Core.Commands.Moderation
     [Category(Categories.Mod)]
     public class KickCommand : BaseCommandModule
     {
-        private readonly IModerationService _moderationService;
-        private readonly ILogger<KickCommand> _logger;
         private readonly IInfractionService _infractionService;
-        public KickCommand(ILogger<KickCommand> logger, IModerationService moderationService, IInfractionService infractionService)
-        {
-            _logger = logger;
-            _moderationService = moderationService;
-            _infractionService = infractionService;
-        }
+        public KickCommand(IInfractionService infractionService) => _infractionService = infractionService;
 
+        
         [Command]
         [RequireFlag(UserFlag.Staff)]
         [RequirePermissions(Permissions.KickMembers)]
@@ -44,7 +37,8 @@ namespace Silk.Core.Commands.Moderation
                     InfractionResult.FailedGuildHeirarchy => "I can't kick that person due to role heiarchy!",
                     InfractionResult.FailedSelfPermissions => "I don't have permission to kick members!", /* In rectrospect, these should never happen, but. */
                     InfractionResult.SucceededWithNotification => $"Kicked {Formatter.Bold($"{user.Username}#{user.Discriminator}")}  (Notified with direct message).",
-                    InfractionResult.SucceededWithoutNotification => $"Kicked {Formatter.Bold($"{user.Username}#{user.Discriminator}")} (Unable to notify with Direct Message)."
+                    InfractionResult.SucceededWithoutNotification => $"Kicked {Formatter.Bold($"{user.Username}#{user.Discriminator}")} (Unable to notify with Direct Message).",
+                    _ => $"Unexpected response: {response}"
                 };
 
                 await ctx.RespondAsync(message);
