+++ src/Silk.Core/Commands/Server/SetMuteCommand.cs
@@ -37,7 +37,7 @@ namespace Silk.Core.Commands.Server
 
             builder.WithContent($"Alright! I'll use {role.Mention} for muting!");
             await ctx.RespondAsync(builder);
-            await _mediator.Send(new UpdateGuildConfigRequest(ctx.Guild.Id) {MuteRoleId = role.Id});
+            await _mediator.Send(new UpdateGuildModConfigRequest(ctx.Guild.Id) {MuteRoleId = role.Id});
 
         }
     }
