+++ src/Silk.Core/EventHandlers/MemberRemoved/MemberRemovedHandler.cs
@@ -17,7 +17,7 @@ namespace Silk.Core.EventHandlers.MemberRemoved
 
         public async Task OnMemberRemoved(DiscordClient c, GuildMemberRemoveEventArgs e)
         {
-            GuildConfig config = await _configService.GetConfigAsync(e.Guild.Id);
+            GuildModConfig config = await _configService.GetModConfigAsync(e.Guild.Id);
             // This should be done in a seperate service //
             if (config.LogMemberLeaves && config.LoggingChannel is not 0)
                 await e.Guild.GetChannel(config.LoggingChannel).SendMessageAsync(GetLeaveEmbed(e));
