+++ src/Silk.Core/EventHandlers/Messages/AutoMod/MessageEditAntiInvite.cs
@@ -1,33 +1,29 @@
 ﻿using System.Threading.Tasks;
 using DSharpPlus;
 using DSharpPlus.EventArgs;
-using Silk.Core.Data.Models;
 using Silk.Core.Services.Data;
-using Silk.Core.Services.Interfaces;
 
 namespace Silk.Core.EventHandlers.Messages.AutoMod
 {
     public class MessageEditAntiInvite
     {
         private readonly ConfigService _configService;
-        private readonly IModerationService _moderationService;
-
-        public MessageEditAntiInvite(IModerationService moderationService, ConfigService configService)
+        private readonly AntiInviteHelper _inviteHelper;
+        public MessageEditAntiInvite(ConfigService configService, AntiInviteHelper inviteHelper)
         {
-            _moderationService = moderationService;
             _configService = configService;
+            _inviteHelper = inviteHelper;
         }
 
         public async Task CheckForInvite(DiscordClient client, MessageUpdateEventArgs args)
         {
             if (args.Channel.IsPrivate) return;
-            GuildConfig config = await _configService.GetConfigAsync(args.Guild.Id);
-
-            bool hasInvite = AntiInviteCore.CheckForInvite(client, args.Message, config, out string invite);
-            bool isBlacklisted = await AntiInviteCore.IsBlacklistedInvite(client, args.Message, config, invite);
+            var config = await _configService.GetModConfigAsync(args.Guild.Id);
+            bool hasInvite = _inviteHelper.CheckForInvite( args.Message, config, out string invite);
+            bool isBlacklisted = await _inviteHelper.IsBlacklistedInvite(args.Message, config, invite);
 
             if (hasInvite && isBlacklisted)
-                await AntiInviteCore.TryAddInviteInfractionAsync(config, args.Message, _moderationService);
+                await _inviteHelper.TryAddInviteInfractionAsync(args.Message);
         }
     }
 }
